LL-4667 Stargate: call getTransactionStatus before attempting to fetch gas estimate (suite)
